% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_idam.R
\docType{data}
\name{agriculture_idam}
\alias{agriculture_idam}
\title{agriculture_idam - Plant production data from the Institute of Sustainable Agricultural and Forestry Development of Amazonas - IDAM (2023).}
\format{
\subsection{`agriculture_idam' A data frame with 1072 rows and 13 columns:}{

\describe{
\item{municipality}{Municipality}
\item{local_unit}{Local Unit of the Institute}
\item{n_beneficiary}{Number of assisted beneficiaries}
\item{planted}{Planted area in hectares (ha)}
\item{harvested}{Harvested area in hectares (ha)}
\item{production}{Production (in the unit of measurement - measure)}
\item{estimated_n_beneficiary}{Estimate of the number of beneficiaries}
\item{estimated_planted}{Estimate of the planted area (ha)}
\item{estimated_harvested}{Estimate of the harvested area (ha)}
\item{estimated_production}{Estimate of production (in the unit of measurement - measure)}
\item{measure}{Unit of Measurement of production}
\item{cultivation}{Cultivation}
\item{year}{Year of cultivation}
}
}
}
\source{
https://www.idam.am.gov.br/biblioteca/relatorio-de-atividades-rat-23-2/
}
\usage{
agriculture_idam
}
\description{
This database, which is the consolidation of data from IDAM's activity reports, provides crop production data for crops that this institute considers a priority.
}
\keyword{datasets}
