% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_manaus.R
\docType{data}
\name{base_manaus}
\alias{base_manaus}
\title{base_manaus - Time series with Extreme Values of Limited Range}
\format{
\subsection{`manaus' A tibble with 144 rows and 5 columns}{

\describe{
\item{year}{The year in which this data was observed}
\item{rh}{Relative Humidity}
\item{dbt}{Dry Bulb Temperature}
\item{wbt}{Wet Bulb Temperature}
\item{month}{The month in which this data was observed}
}
}
}
\source{
Source information to be provided later
}
\usage{
base_manaus
}
\description{
A time series dataset consisting of the minimum relative humidity observed in the city of Manaus from January 2009 to December 2020. This dataset is a sample of size n = 144 monthly blocks.
}
\keyword{datasets}
