% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srl_muni.R
\docType{data}
\name{srl_muni}
\alias{srl_muni}
\title{srl_muni - School Read Levels - Physical literacy and reading performance of amazonian schoolchildren}
\format{
\subsection{`srl_muni' A data frame with 100 rows and 17 columns:}{

\describe{
\item{school}{School of observation}
\item{sex}{Sex - 1 for Masculine and 2 for Feminine}
\item{age}{Age in years}
\item{weight_kg}{Weight in kilograms (kg)}
\item{height_m}{Height in meters (m)}
\item{bmi}{Body Mass Index (BMI)}
\item{bmi_status}{Body Mass Index Status - A categorical description of the BMI}
\item{phv}{Peak Height Velocity (PHV) is a continuous variable, measured by age (in years) relative to the peak height in the period where maximum rate of growth occurs and this is an indication of maturity}
\item{phv_state}{Categorical description for Peak Height Velocity (PHV)}
\item{eses}{Exercise Self-Efficacy Scale measures how confident the students are with regard to carrying out regular physical activities and exercise, measured by a psychometric scale}
\item{pmc}{Perceived Motor Competence (PMC) assesses the child's self-perception regarding their ability to perform motor movements. Measured by a psychometric scale}
\item{kupa}{Knowledge and Understanding of Physical Activity (KUPA) - Evaluates knowledge regarding the benefits, safety, and engagement in physical activity. Measured by scores obtained in objective questions}
\item{sb}{Sedentary Behavior - Evalutes the exposition time to sedentary behaviors. Measured by psychometric scale}
\item{pa}{Pyshical Activity (PA) - Level of physical activity ian children, based on the frequency and duration of routine daily activities. Measured by score on the physical activity frequency questions}
\item{mc}{Motor Competence - Supine To Stand (in seconds) - Measures the time (in seconds) it takes for the child to stand up from the floor, from a supine position to a standing position}
\item{tde}{Total Score on the TDE test - Measures reading level based on time and accuracy in word reading}
\item{tde_categorical}{Categorical description of the TDE Score}
}
}
}
\source{
ALMEIDA, Thiago da Cruz de. Physical Literacy e desempenho em leitura de escolares amazônicos: um estudo de associação. 2024. 104 f. Dissertação (Mestrado em Educação) - Universidade Federal do Amazonas, Manaus, 2024. Disponível em: https://tede.ufam.edu.br/handle/tede/10627.
}
\usage{
srl_muni
}
\description{
This is a dataset from an association, cross-sectional, quantitative study carried out with Amazonian schoolchildren (n=100; aged between 10 and 13 years) from 3 public schools in the municipality of Borba (Amazonas).
}
\keyword{datasets}
