% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aids_amazonas.R
\docType{data}
\name{aids_amazonas}
\alias{aids_amazonas}
\title{aids_amazonas - A dataset of AIDS occurrences from 2011 to 2023 in Amazonas}
\format{
\subsection{`aids_amazonas' A data frame with 1612 rows and 5 columns:}{

\describe{
\item{code_muni}{Municipality code}
\item{name_muni}{Municipality name}
\item{cases}{Case count}
\item{gender}{Gender of the infected individual - M for Male and F for female}
\item{year}{Year of observation}
}
}
}
\source{
https://indicadores.aids.gov.br/
}
\usage{
aids_amazonas
}
\description{
This dataset contains data of the AIDS occurrences in each municipality from Amazonas.
}
\keyword{datasets}
