% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_amazonas.R
\docType{data}
\name{malaria_amazonas}
\alias{malaria_amazonas}
\title{malaria_amazonas - An integrated dataset of malaria notifications in the state of Amazonas}
\format{
\subsection{`malaria_amazonas' A data frame with 551633 rows and 35 columns:}{

\describe{
\item{notification_hr}{Health region of notification}
\item{notification_county}{County of notification}
\item{infection_state}{Probable state where the patient was infected}
\item{infection_hr}{Probable health region where the patient was infected}
\item{infection_county}{Probable county where the patient was infected}
\item{home_hr}{Health region of residence of the patient}
\item{home_county}{County of residence of the patient}
\item{exam_interval}{Time interval between notification and examination}
\item{treatment_interval}{Time interval between examination and beginning of treatment}
\item{notification_interval}{Time interval between symptom and notification}
\item{notification_month}{Month in which the notification was recorded}
\item{notification_year}{Year in which the notification was recorded}
\item{symptom_month}{Month in which the patient felt the first symptoms of malaria}
\item{symptom_year}{Year in which the patient felt the first symptoms of malaria}
\item{exam_month}{Month in which the examination was performed}
\item{exam_year}{Year in which the examination was performed}
\item{treatment_month}{Month in which the treatment started}
\item{treatment_year}{Year in which the treatment started}
\item{migration}{Health region of residence different than that of notification}
\item{autochthonous_case}{Health region of infection identical to that of residence}
\item{exam_result}{Result of examination}
\item{detection_type}{Type of detection}
\item{exam_type}{Type of examination}
\item{symptom}{Indicates if the patient felt a symptom}
\item{hemiparasite}{The result of the examination for other hemiparasites}
\item{previous_treatment}{Previous treatment for P. vivax or P.falciparum}
\item{occupation}{Main activity in the last 15 days}
\item{education_level}{Level of education of the patient}
\item{age}{Interval of the age of  the patient}
\item{race}{Race/color of the patient}
\item{gender}{Gender of the patient}
\item{pregnancy}{Indicates pregnancy and the gestational age}
\item{crosses}{Indicates the amount of parasitemia in crosses}
\item{cvl_case}{Indicates the existence of cases of canine visceral leishmaniasis}
\item{scheme}{Indicates the treatment scheme employed}
\item{qty_parasites}{Indicates the number of parasites per mm^3}
}
}
}
\source{
https://www.synapse.org/Synapse:syn21552203/files
}
\usage{
malaria_amazonas
}
\description{
This dataset contains medical records of patients who underwent malaria screening tests in the Legal Amazon. The data in this dataset were filtered to include only the cases reported in the state of Amazonas and for patients who reside specifically in the state of Amazonas between 2014 and 2019.
}
\examples{
\donttest{
# Bar plot of malaria tests results (2019)
library(dplyr)
library(ggplot2)

malaria_amazonas \%>\%
  filter(exam_year == 2019) \%>\%
  count(exam_result) \%>\%
  ggplot(aes(x = reorder(exam_result, -n), y = n)) +
  geom_bar(stat = "identity") +
  geom_text(
    aes(label = n),
    vjust = -0.5
  ) +
  theme_minimal() +
  labs(
    title = "Results of malaria tests conducted in 2019",
    subtitle = "Documented cases in Amazonas and of patients residing in the state",
    x = "Result",
    y = "Count"
  )



}
}
\keyword{datasets}
